+++ __tests__/hooks/useEtherspotTransactions.test.js
@@ -299,17 +299,31 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isSending).toBe(false);
 
+    // Mocking the timeout during the send operation
+    jest.useFakeTimers();
+
     let sendPromise;
     act(() => {
       sendPromise = result.current.send();
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    // Simulate 30 seconds timeout
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(
+      () => {
+        expect(result.current.isSending).toBe(false);
+      },
+      { timeout: 31000 }
+    ); // Increase timeout slightly to cover the full timeout
 
     const sent = await sendPromise;
     expect(result.current.containsSendingError).toBe(false);
     expect(sent[0].sentBatches[0].userOpHash).toBe('0x7c');
+
+    jest.useRealTimers();
   });
 
   it('estimates and sends single grouped batches without calling estimate', async () => {
@@ -350,18 +364,27 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isSending).toBe(false);
 
+    jest.useFakeTimers();
+
     let sendPromise;
     act(() => {
       sendPromise = result.current.send();
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent = await sendPromise;
     expect(result.current.containsSendingError).toBe(false);
     expect(sent[0].estimatedBatches[0].cost.toString()).toBe('350000000000000');
     expect(sent[0].sentBatches[0].userOpHash).toBe('0x7c');
+
+    jest.useRealTimers();
   });
 
   it('estimates and sends multiple grouped batches with skipped and with no batches', async () => {
@@ -420,13 +443,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate();
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(false);
@@ -449,7 +479,12 @@ describe('useEtherspotTransactions()', () => {
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent = await sendPromise;
     expect(result.current.containsSendingError).toBe(false);
@@ -457,6 +492,8 @@ describe('useEtherspotTransactions()', () => {
     expect(sent[0].sentBatches[1].userOpHash).toBe('0x7d');
     expect(sent[1].sentBatches.length).toBe(0); // has skip prop
     expect(sent[2].sentBatches[0].userOpHash).toBe('0x46');
+
+    jest.useRealTimers();
   });
 
   it('estimates and sends multiple grouped batches with paymaster', async () => {
@@ -502,13 +539,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate();
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(false);
@@ -527,12 +571,19 @@ describe('useEtherspotTransactions()', () => {
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent = await sendPromise;
     expect(result.current.containsSendingError).toBe(false);
     expect(sent[0].sentBatches[0].userOpHash).toBe('0x7c');
     expect(sent[1].sentBatches[0].userOpHash).toBe('0x46');
+
+    jest.useRealTimers();
   });
 
   it('estimates and sends multiple grouped batches with matching chain IDs', async () => {
@@ -590,13 +641,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate();
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(false);
@@ -621,7 +679,12 @@ describe('useEtherspotTransactions()', () => {
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent = await sendPromise;
     expect(result.current.containsSendingError).toBe(false);
@@ -629,6 +692,8 @@ describe('useEtherspotTransactions()', () => {
     expect(sent[0].sentBatches[1].userOpHash).toBe('0x7e');
     expect(sent[1].sentBatches[0].userOpHash).toBe('0x46');
     expect(sent[2].sentBatches[0].userOpHash).toBe('0x7d');
+
+    jest.useRealTimers();
   });
 
   it('estimates and calls onEstimated for each batch group', async () => {
@@ -687,13 +752,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate();
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(false);
@@ -705,6 +777,8 @@ describe('useEtherspotTransactions()', () => {
     expect(onEstimated2.mock.calls[0][0]).toStrictEqual(
       estimated[1].estimatedBatches
     );
+
+    jest.useRealTimers();
   });
 
   it('sends and calls onSent for each batch group', async () => {
@@ -763,13 +837,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isSending).toBe(false);
 
+    jest.useFakeTimers();
+
     let sendPromise;
     act(() => {
       sendPromise = result.current.send();
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent = await sendPromise;
 
@@ -777,6 +858,8 @@ describe('useEtherspotTransactions()', () => {
     expect(onSent2).toBeCalledTimes(1);
     expect(onSent1.mock.calls[0][0]).toStrictEqual(sent[0].sentBatches);
     expect(onSent2.mock.calls[0][0]).toStrictEqual(sent[1].sentBatches);
+
+    jest.useRealTimers();
   });
 
   it('estimates and returns error messages for each batch group', async () => {
@@ -832,13 +915,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate();
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(true);
@@ -854,6 +944,8 @@ describe('useEtherspotTransactions()', () => {
     expect(onEstimated2.mock.calls[0][0]).toStrictEqual(
       estimated[1].estimatedBatches
     );
+
+    jest.useRealTimers();
   });
 
   it('estimates successfully and returns error messages on send for each batch group', async () => {
@@ -905,13 +997,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate();
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(false);
@@ -930,7 +1029,12 @@ describe('useEtherspotTransactions()', () => {
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent = await sendPromise;
 
@@ -943,6 +1047,8 @@ describe('useEtherspotTransactions()', () => {
     );
     expect(onSent1.mock.calls[0][0]).toStrictEqual(sent[0].sentBatches);
     expect(onSent2.mock.calls[0][0]).toStrictEqual(sent[1].sentBatches);
+
+    jest.useRealTimers();
   });
 
   it('estimates valid and returns error messages for invalid batch group by ID', async () => {
@@ -1006,13 +1112,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isEstimating).toBe(false);
 
+    jest.useFakeTimers();
+
     let estimatePromise;
     act(() => {
       estimatePromise = result.current.estimate(['test-id-1']);
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated1 = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(true);
@@ -1029,7 +1142,9 @@ describe('useEtherspotTransactions()', () => {
     });
 
     await waitFor(() => expect(result.current.isEstimating).toBe(true));
-    await waitFor(() => expect(result.current.isEstimating).toBe(false));
+    await waitFor(() => expect(result.current.isEstimating).toBe(false), {
+      timeout: 31000,
+    });
 
     const estimated2 = await estimatePromise;
     expect(result.current.containsEstimatingError).toBe(false);
@@ -1046,6 +1161,8 @@ describe('useEtherspotTransactions()', () => {
     expect(onEstimated3.mock.calls[0][0]).toStrictEqual(
       estimated2[1].estimatedBatches
     );
+
+    jest.useRealTimers();
   });
 
   it('sends valid and returns error messages for invalid batch group by ID', async () => {
@@ -1109,13 +1226,20 @@ describe('useEtherspotTransactions()', () => {
 
     expect(result.current.isSending).toBe(false);
 
+    jest.useFakeTimers();
+
     let sendPromise;
     act(() => {
       sendPromise = result.current.send(['test-id-1']);
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+
+    jest.advanceTimersByTime(30000);
+
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent1 = await sendPromise;
     expect(sent1.length).toBe(1);
@@ -1130,7 +1254,9 @@ describe('useEtherspotTransactions()', () => {
     });
 
     await waitFor(() => expect(result.current.isSending).toBe(true));
-    await waitFor(() => expect(result.current.isSending).toBe(false));
+    await waitFor(() => expect(result.current.isSending).toBe(false), {
+      timeout: 31000,
+    });
 
     const sent2 = await sendPromise;
     expect(sent2.length).toBe(2);
@@ -1139,5 +1265,7 @@ describe('useEtherspotTransactions()', () => {
     expect(sent2[1].sentBatches[0].userOpHash).toBe('0x47');
     expect(onSent2.mock.calls[0][0]).toStrictEqual(sent2[0].sentBatches);
     expect(onSent3.mock.calls[0][0]).toStrictEqual(sent2[1].sentBatches);
+
+    jest.useRealTimers();
   });
 });