+++ src/providers/EtherspotTransactionKitContextProvider.tsx
@@ -211,7 +211,45 @@ const EtherspotTransactionKitContextProvider = ({
             const userOpHash = await etherspotModulaSdk.send(
               estimatedBatch.userOp
             );
-            sentBatches.push({ ...estimatedBatch, userOpHash });
+
+            // get transaction hash or userOp receipt...
+            let userOpsReceipt;
+            const timeout = Date.now() + 30 * 1000; // 30 seconds timeout
+
+            try {
+              while (!userOpsReceipt && Date.now() < timeout) {
+                await new Promise<void>((resolve) => {
+                  setTimeout(resolve, 2000);
+                }); // Retry every 2 sec
+
+                try {
+                  userOpsReceipt =
+                    await etherspotModulaSdk.getUserOpReceipt(userOpHash);
+                } catch (error) {
+                  console.error(
+                    'Error fetching transaction hash. Please check if the transaction has gone through, or try to send the transaction again:',
+                    error
+                  );
+                }
+              }
+
+              if (!userOpsReceipt) {
+                console.warn(
+                  'Failed to get the transaction hash within 30 seconds. Please check if the transaction has gone through, or try to send the transaction again.'
+                );
+              }
+            } catch (e) {
+              console.error(
+                'Unexpected error while fetching the transaction hash. Please check if the transaction has gone through, or try to send the transaction again:',
+                e
+              );
+            }
+
+            sentBatches.push({
+              ...estimatedBatch,
+              userOpHash,
+              transactionHash: userOpsReceipt,
+            });
           } catch (e) {
             const errorMessage = parseEtherspotErrorMessage(
               e,
@@ -240,11 +278,47 @@ const EtherspotTransactionKitContextProvider = ({
     return result;
   };
 
+  const getTransactionHash = async (
+    userOpHash: string,
+    batchId: number,
+    timeout: number = 30 * 1000,
+    retryInterval: number = 2000
+  ): Promise<string | null> => {
+    const etherspotModulaSdk = await getSdk(batchId);
+
+    let transactionHash = null;
+    const timeoutTotal = Date.now() + timeout; // 30 seconds timeout
+
+    while (!transactionHash && Date.now() < timeoutTotal) {
+      await new Promise<void>((resolve) => {
+        setTimeout(resolve, retryInterval);
+      }); // Retry every 2 sec
+
+      try {
+        transactionHash = await etherspotModulaSdk.getUserOpReceipt(userOpHash);
+      } catch (error) {
+        console.error(
+          'Error fetching transaction hash. Please try again:',
+          error
+        );
+      }
+    }
+
+    if (!transactionHash) {
+      console.warn(
+        'Failed to get the transaction hash within 30 seconds. Please try again'
+      );
+    }
+
+    return transactionHash;
+  };
+
   const contextData = useMemo(
     () => ({
       batches: getObjectSortedByKeys(groupedBatchesPerId),
       estimate,
       send,
+      getTransactionHash,
       chainId,
       isEstimating,
       isSending,